Move PeerCommunicator and BackgroundListener into its own module peer_communicator
