+++ exasol_advanced_analytics_framework/udf_communication/peer_communicator/__init__.py
@@ -1,0 +1,1 @@
+from .peer_communicator import PeerCommunicator
