+++ exasol_advanced_analytics_framework/udf_communication/peer_communicator/frontend_peer_state.py
@@ -1,14 +1,12 @@
 import contextlib
-import time
 from typing import Optional, Generator, List
 
 import zmq
 
 from exasol_advanced_analytics_framework.udf_communication.connection_info import ConnectionInfo
-from exasol_advanced_analytics_framework.udf_communication.get_peer_receive_socket_name import \
+from exasol_advanced_analytics_framework.udf_communication.peer_communicator.get_peer_receive_socket_name import \
     get_peer_receive_socket_name
-from exasol_advanced_analytics_framework.udf_communication.messages import Message, WeAreReadyToReceiveMessage, \
-    AreYouReadyToReceiveMessage, PayloadMessage, PeerIsReadyToReceiveMessage
+from exasol_advanced_analytics_framework.udf_communication.messages import PayloadMessage
 from exasol_advanced_analytics_framework.udf_communication.peer import Peer
 from exasol_advanced_analytics_framework.udf_communication.serialization import serialize_message
