+++ exasol_advanced_analytics_framework/udf_communication/peer_communicator/get_peer_receive_socket_name.py
