+++ setup.py
@@ -11,6 +11,7 @@ packages = \
  'exasol_advanced_analytics_framework.query_result',
  'exasol_advanced_analytics_framework.testing',
  'exasol_advanced_analytics_framework.udf_communication',
+ 'exasol_advanced_analytics_framework.udf_communication.peer_communicator',
  'exasol_advanced_analytics_framework.udf_framework']
 
 package_data = \
